refactoring frontend file structure
