+++ frontend/src/Wishlist/ListRecipe.css
